requirements.txt
python-gedcom
python-docx